// src/app/Model/production.ts
export class FlaskListSearch {
  PartNo: string = "";
  FlaskId: string = "";
  FLength: number = 0;
  FWidth: number = 0;
  FUpHeight: number = 0;
  FDHeight: number = 0;
  FMHeight: number = 0;
  FlaskLocation: string = "";
  Choice: boolean = false;
  Status: string = '';
  Usage: number = 0;
}
export class FlaskList {
  PartNo: string = "";
  FlaskId: string = "";
  FLength: number = 0;
  FWidth: number = 0;
  FUpHeight: number = 0;
  FDHeight: number = 0;
  FMHeight: number = 0;
  FlaskLocation: string = "";
  Choice: boolean = false;
}

export class QualityRecord {
  TA001: string = "";
  TA002: string = "";
  QualityIssuseDate: string = "";
  PartNo: string = "";
  ProblemRemark: string = "";

}
export class WIPProcess {
  PartNo: string = "";
  WIPQTY: number = 0;
  WIPProcessCode: string = "";
  WIPProcessName: string = "";
  WIPProcessStartDate: string = "";
  Flask: string = "";
  TA002: string = "";
  Create_ID: number = 0;
  LastProceeID: number = 0;
  HistoryString: string = "";
}
// ...
export class FlaskUsageList {
  PartNO: string = "";
  Weight: number = 0;
  FlaskUse: number = 0;
}
// ...
export class SchedulingTable {
  scheduleindex: number = 0;
  PartNo: string = "";
  OrderType: string = "";
  OrderNo: string = "";
  OrderSeq: string = "";
  Weight: number = 0;
  ETD: string = "";
  FlaskNo: string = "";
  Pattern: string = "";
  Molding: string = "";
  MoldingLine: string = "";
  MoldingWIPProcessCode: string = '';
  Asembling: string = "";
  AsemblingLine: string = "";
  AsemblingWIPProcessCode: string = '';
  Pouring: string = "";
  DeFlask: string = "";
  StressRelease: string = "";
  DeBurring: string = "";
  OutSourcing: string = "";
  ProductionOrderHead: string = '';
  ProductionOrder: string = "";
  EstimateFinDate: string = "";
  Remark: string = "";
  PatternLocation = "";
  FlaskLocation = "";
  PatternLength: number = 0;
  PatternUHeight: number = 0;
  PatternDHeight: number = 0;
  PatternWidth: number = 0;
  SelectWeekProduction: string = '';
  OrderHeadCode: string = '';
  OrderCode: string = '';
  OrderDate: string = '';
  BillingDate: string = "";
  OrderSeries: string = '';
}
// ...
export class ProductionOrder {
  PartNo: string = "";
  ProductionOrderNo: string = "";
  OrderQTY: number = 0;
  FindQTY: number = 0;
}

export class CusOrder {
  PartNo: string = "";
  CusOrderNo: string = "";
  OrderQTY: number = 0;
  FindQTY: number = 0;
  PONum: string = '';
  FinDate: string = '';
  ShipLocation: string = '';
  TD001: string = '';
  TD002: string = '';
  D003: string = '';
}

export class StockRecord {
  PartNo: string = "";
  StockQty: number = 0;
  StockInDate: string = "";
  Location: string = '';
}

export class PurchaseRecord {
  PartNo: string = "";
  PurchaseNo: string = "";
  PurchaseQTY: number = 0;
  FindQTY
}

export class CastingPurchase {
  PartNo: string = '';
  PartDesc: string = '';
  PurQTY: number = 0;
  FinQTY: number = 0;
  OrderFinDate: string = '';
  SupplierName: string = '';
  PurchaseDate: string = '';
  PurchaseFinishedList: Array<PurchaseFinished> = new Array<PurchaseFinished>();
}
export class PurchaseFinished {
  PartNo: string = '';
  FinQTY: number = 0;
  FinDate: string = '';
}
export class POstatus {
  PartNO: string = "";
  PartDesc: string = "";
  CusOrder: string = "";
  OrderQTY: number = 0;
  StockQTY: number = 0;
  QTYString: string = "";
  OrderFinDate: string = "";
  Weight: number = 0;
  CusName: string = "";
  CusNO: string = "";
  ShipLoc: string = "";
  OrderStop: string = "";
  TA026: string = "";
  TA027: string = "";
  TA028: string = "";
  Remark: string = "";
  TA002: string = "";
  TD016: string = "";
  TC027: string = "";
  MoldingLine: string = "";
  AsemblingLine: string = "";
  WIPQTY: number = 0;
  WIPProcessList: Array<WIPProcess> = new Array<WIPProcess>();
  PlanProductionQTY: number = 0;
  runnungqty: number = 0;
  BatchQTY: number = 0;
  MoldingTime: number = 0;
  AsemblingTime: number = 0;
  PouringTime: number = 0;
  DeflaskTime: number = 0;
  StressFuranceTime: number = 0;
  DeburringTime: number = 0;
  flaskMatches: Array<FlaskList> = new Array<FlaskList>();
  PatternLocation: string = "";
  PatternUHeight: number = 0;
  PatternDHeight: number = 0;
  PatternWidth: number = 0;
  PatternLength: number = 0;
  QualityRecordList: Array<QualityRecord> = new Array<QualityRecord>();
  OutSourceList: Array<WIPProcess> = new Array<WIPProcess>();
  CastingPurchaseList: Array<CastingPurchase> = new Array<CastingPurchase>();
}